requirements.txt
playwright==1.41.0
pandas==2.2.0
openpyxl==3.1.2
python-dotenv==1.0.0
colorama==0.4.6
